python/requirements.txt
# google cloud lib
google-cloud-bigquery

# 
psycopg2
